Integrate feedback into serialization
